refactor(coding): own coding work shell implementation
