+++ src/loushang/work/__init__.py
@@ -1,4 +1,3 @@
-from loushang.work.coding import CodingWorkShell
 from loushang.work.event_log import (
     EventLogBackend,
     EventLogEntry,
@@ -47,3 +46,11 @@ __all__ = [
     "project_agent_event_to_work_events",
     "project_work_plan_runs",
 ]
+
+
+def __getattr__(name: str) -> object:
+    if name == "CodingWorkShell":
+        from loushang.coding.work_shell import CodingWorkShell
+
+        return CodingWorkShell
+    raise AttributeError(f"module 'loushang.work' has no attribute {name!r}")
