+++ tests/coding/test_work_shell_adapter.py
@@ -6,3 +6,4 @@ def test_coding_work_shell_adapter_exposes_coding_owned_entrypoint() -> None:
     from loushang.work import CodingWorkShell as CompatCodingWorkShell
 
     assert CodingWorkShell is CompatCodingWorkShell
+    assert CodingWorkShell.__module__ == "loushang.coding.work_shell"
